refactor(Historisation): bouger les signals dans le model
